# 中括弧'{{'検索処理をループからReadOnlySpan<byte>.IndexOf()に変更