``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1165 (21H1/May2021Update)
AMD Ryzen 7 5800U with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.100-preview.7.21379.14
  [Host]   : .NET 6.0.0 (6.0.21.37719), X64 RyuJIT
  .NET 6.0 : .NET 6.0.0 (6.0.21.37719), X64 RyuJIT

Job=.NET 6.0  Runtime=.NET 6.0  

```
| Method |      Mean |     Error |    StdDev | Allocated |
|------- |----------:|----------:|----------:|----------:|
|   Main | 17.789 ns | 0.0744 ns | 0.0659 ns |         - |
|     PR |  8.494 ns | 0.0106 ns | 0.0094 ns |         - |