@finphie 2021-09-05
'}}'検索処理にもIndexOfを使うべきだと思うが、別のPRで対応する。
先頭末尾スペース削除処理には有効か分からない。（スペースは通常一文字のため）